add BucketsByIndices & TotalBucketCount refactor delta

+++ blockindex/liquidstaking_dirty.go
@@ -7,6 +7,7 @@ package blockindex
 
 import (
 	"math/big"
+	"sync"
 	"time"
 
 	"github.com/iotexproject/iotex-core/db/batch"
@@ -24,46 +25,9 @@ type (
 		delta      *liquidStakingDelta // delta for cache to store buckets of current block
 		batch      batch.KVStoreBatch  // batch for db to store buckets of current block
 		tokenOwner map[uint64]string
-	}
-
-	liquidStakingDelta struct {
-		*liquidStakingCache // easy to query buckets
-
-		updatedBucketType map[uint64]*BucketType
-		updatedBucketInfo map[uint64]*BucketInfo
-		deletedBucketInfo map[uint64]bool
+		once       sync.Once
 	}
 )
-
-func newLiquidStakingDelta() *liquidStakingDelta {
-	return &liquidStakingDelta{
-		liquidStakingCache: newLiquidStakingCache(),
-		updatedBucketType:  make(map[uint64]*BucketType),
-		updatedBucketInfo:  make(map[uint64]*BucketInfo),
-		deletedBucketInfo:  make(map[uint64]bool),
-	}
-}
-
-func (s *liquidStakingDelta) putBucketType(id uint64, bt *BucketType) {
-	s.liquidStakingCache.putBucketType(id, bt)
-	s.updatedBucketType[id] = bt
-}
-
-func (s *liquidStakingDelta) putBucketInfo(id uint64, bi *BucketInfo) {
-	s.liquidStakingCache.putBucketInfo(id, bi)
-	s.updatedBucketInfo[id] = bi
-	if s.deletedBucketInfo[id] {
-		delete(s.deletedBucketInfo, id)
-	}
-}
-
-func (s *liquidStakingDelta) deleteBucketInfo(id uint64) {
-	s.liquidStakingCache.deleteBucketInfo(id)
-	s.deletedBucketInfo[id] = true
-	if _, ok := s.updatedBucketInfo[id]; ok {
-		delete(s.updatedBucketInfo, id)
-	}
-}
 
 func newLiquidStakingDirty(clean *liquidStakingCache) *liquidStakingDirty {
 	return &liquidStakingDirty{
@@ -75,37 +39,51 @@ func newLiquidStakingDirty(clean *liquidStakingCache) *liquidStakingDirty {
 }
 
 func (s *liquidStakingCache) merge(delta *liquidStakingDelta) error {
-	for id, bt := range delta.updatedBucketType {
-		s.putBucketType(id, bt)
+	for id, state := range delta.bucketTypeDeltaState {
+		if state == deltaStateAdded || state == deltaStateModified {
+			s.putBucketType(id, delta.mustGetBucketType(id))
+		}
 	}
-	for id, bi := range delta.updatedBucketInfo {
-		s.putBucketInfo(id, bi)
-	}
-	for id := range delta.deletedBucketInfo {
-		s.deleteBucketInfo(id)
+	for id, state := range delta.bucketInfoDeltaState {
+		if state == deltaStateAdded || state == deltaStateModified {
+			s.putBucketInfo(id, delta.mustGetBucketInfo(id))
+		} else if state == deltaStateRemoved {
+			s.deleteBucketInfo(id)
+		}
 	}
 	s.putHeight(delta.getHeight())
+	s.putTotalBucketCount(s.getTotalBucketCount() + delta.addedBucketCnt())
 	return nil
 }
 
 func (s *liquidStakingDirty) putHeight(h uint64) {
-	s.batch.Put(_liquidStakingHeightNS, _liquidStakingHeightKey, byteutil.Uint64ToBytesBigEndian(h), "failed to put height")
+	s.batch.Put(_liquidStakingNS, _liquidStakingHeightKey, byteutil.Uint64ToBytesBigEndian(h), "failed to put height")
 	s.delta.putHeight(h)
 }
 
-func (s *liquidStakingDirty) putBucketType(id uint64, bt *BucketType) {
+func (s *liquidStakingDirty) addBucketType(id uint64, bt *BucketType) error {
 	s.batch.Put(_liquidStakingBucketTypeNS, byteutil.Uint64ToBytesBigEndian(id), bt.serialize(), "failed to put bucket type")
-	s.delta.putBucketType(id, bt)
+	return s.delta.addBucketType(id, bt)
 }
 
-func (s *liquidStakingDirty) putBucketInfo(id uint64, bi *BucketInfo) {
+func (s *liquidStakingDirty) updateBucketType(id uint64, bt *BucketType) error {
+	s.batch.Put(_liquidStakingBucketTypeNS, byteutil.Uint64ToBytesBigEndian(id), bt.serialize(), "failed to put bucket type")
+	return s.delta.updateBucketType(id, bt)
+}
+
+func (s *liquidStakingDirty) addBucketInfo(id uint64, bi *BucketInfo) error {
 	s.batch.Put(_liquidStakingBucketInfoNS, byteutil.Uint64ToBytesBigEndian(id), bi.serialize(), "failed to put bucket info")
-	s.delta.putBucketInfo(id, bi)
+	return s.delta.addBucketInfo(id, bi)
 }
 
-func (s *liquidStakingDirty) burnBucket(id uint64) {
+func (s *liquidStakingDirty) updateBucketInfo(id uint64, bi *BucketInfo) error {
+	s.batch.Put(_liquidStakingBucketInfoNS, byteutil.Uint64ToBytesBigEndian(id), bi.serialize(), "failed to put bucket info")
+	return s.delta.updateBucketInfo(id, bi)
+}
+
+func (s *liquidStakingDirty) burnBucket(id uint64) error {
 	s.batch.Delete(_liquidStakingBucketInfoNS, byteutil.Uint64ToBytesBigEndian(id), "failed to delete bucket info")
-	s.delta.deleteBucketInfo(id)
+	return s.delta.deleteBucketInfo(id)
 }
 
 func (s *liquidStakingDirty) getBucketTypeIndex(amount *big.Int, duration time.Duration) (uint64, bool) {
@@ -141,7 +119,7 @@ func (s *liquidStakingDirty) getBucketType(id uint64) (*BucketType, bool) {
 }
 
 func (s *liquidStakingDirty) getBucketInfo(id uint64) (*BucketInfo, bool) {
-	if s.delta.deletedBucketInfo[id] {
+	if s.delta.isBucketDeleted(id) {
 		return nil, false
 	}
 	bi, ok := s.delta.getBucketInfo(id)
@@ -150,4 +128,16 @@ func (s *liquidStakingDirty) getBucketInfo(id uint64) (*BucketInfo, bool) {
 	}
 	bi, ok = s.clean.getBucketInfo(id)
 	return bi, ok
+}
+
+func (s *liquidStakingDirty) finalizeBatch() batch.KVStoreBatch {
+	s.once.Do(func() {
+		total := s.clean.getTotalBucketCount() + s.delta.addedBucketCnt()
+		s.batch.Put(_liquidStakingNS, _liquidStakingTotalBucketCountKey, byteutil.Uint64ToBytesBigEndian(total), "failed to put total bucket count")
+	})
+	return s.batch
+}
+
+func (s *liquidStakingDirty) finalize() (batch.KVStoreBatch, *liquidStakingDelta) {
+	return s.finalizeBatch(), s.delta
 }
